Extracted the conversions in validateAddress Created a new utilities module for all future conversions.
