+++ index.js
@@ -3,7 +3,7 @@ var config = require('./config/config');
 var rp = require('request-promise');
 var restify = require('restify');
 var parser = require('xml2json');
-var _ = require('lodash');
+var converters = require('./converters');
 var options = {
   'auth': {
     'sendImmediately': true
@@ -45,21 +45,9 @@ server.listen(config.port, function () {
 });
 
 function validateAddress(req, res, next) {
-  //res.send('Not Yet Implemented.');
-  //res.send(req.body);
-  var address = {
-    validateLocation: {
-      location: {
-        address1: { $t: req.body.addressLine1 },
-        address2: { $t: req.body.addressLine2 },
-        community: { $t: req.body.community },
-        state: { $t: req.body.state },
-        postalcode: { $t: req.body.postalCode },
-        type: { $t: 'ADDRESS' }
-      }
-    }
-  };
-  var xml = parser.toXml(address);
+  var converter = converters.validateAddress;
+
+  var xml = converter.createXmlString(req.body);
   rp.post(config.dash.url + 'validatelocation', {
     auth: options.auth,
     body: xml,
@@ -71,34 +59,8 @@ function validateAddress(req, res, next) {
       ]
   })
   .then(function (response) {
-    var json = parser.toJson(response, { object: true });
-    var location = json['ns2:validateLocationResponse'].Location; 
-    console.log(location);
-
-    var validatedAddress = {
-      //addressId: Unknown. Not in the Bandwidth response
-      addressLine1: location.address1,
-      addressLine2: _.isEmpty(location.address2) ? '' : location.address2,
-      houseNumber: location.legacydata.housenumber,
-      prefixDirectional: location.legacydata.predirectional,
-      streetName: location.legacydata.streetname,
-      //postDirectional: Unknown. Not in the Bandwidth response
-      //streetSuffix: Unknown. Not in the Bandwidth response
-      community: location.community,
-      state: location.state,
-      //unitType: Unknown. Not in the Bandwidth response
-      //unitTypeValue: Unknown. Not in the Bandwidth response
-      longitude: location.longitude,
-      latitude: location.latitude,
-      postalCode: location.postalcode,
-      zipPlusFour: location.plusfour,
-      //description: Unknown. Description found in response: "Location is geocoded"
-      addressStatus: location.status.code,
-      //createdOn: Unknown. activatedtime/updatetime found in response
-      //modifiedOn: Unknown. activatedtime/updatetime found in response
-    };
-
-    res.json(validatedAddress);
+    var obj = converter.createJsObject(response);
+    res.json(obj);
     next();
   })
   .catch(function (err) {
